complain management system
The main conceptual idea of the Complaint Management System is to register the complaints of the User and to process them by following some steps. Here whenever a User has got any complaint related to the anything then they can make online complaints. Their complaints are being taken care by the staff. The registered complaints of the User, gets forwarded to the Staff, then after the correction of the fault, the line-man again makes a call back to the staff and they make it confirm to the User. 
Our project is an online complaint management system where the issues of the customers can be registered online and resolved by the different levels of staff. Also flexibility is provided to the customers can easily resolve their issues by communicating with staff over internet. Complaint management is a website that acts as a bridge between customer and Pondicherry tourism department in which customer directly register their complaint via internet. Lack of paper movements provides complaint management operations a speed which was never envisaged in manual mode at all. Website allows customer to register complaint and automatically schedules and prompts operators to source complaint to concerned departments.